[artf780788]: Create ApduServiceInfo object with valid offHost Info
Whenever application calls
getRouteDestinationForPreferredPaymentService API,
a new ApduserviceInfo object is created from already
available preferredPaymentService. For offHost services
non null offHostName to be used for object construction.
But this parameter is getting set as null for OffHost/OnHost.
This causes application to wrongly find it as Host application.
To avoid this problem, valid offHost information is updated.
Test:
1. Install offhost payment app,
2. Select the app in Tap & Pay.
3. call getRouteDestinationForPreferredPaymentService